refactor: move built-in reporter modules
