+++ src/builtin-reporters/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Built-in reporter exports.
+ */
+
+export { rawJsonReporter } from "./raw-json/index.ts";
+export { summaryReporter } from "./summary/index.ts";
