+++ src/builtin-reporters/raw-json/serializer.test.ts
