+++ src/builtin-reporters/summary/diagnostic-items.ts
