+++ src/builtin-reporters/summary/diagnostics-printer.ts
