+++ src/builtin-reporters/summary/finding-printer.ts
