+++ src/builtin-reporters/summary/formatter.ts
