+++ src/builtin-reporters/summary/inspector-result-printer.ts
