+++ src/builtin-reporters/summary/markup-printer/element-printer.ts
