+++ src/builtin-reporters/summary/markup-printer/list-printer.ts
