+++ src/builtin-reporters/summary/markup-printer/paragraph-printer.ts
