+++ src/builtin-reporters/summary/markup-printer/spacing.ts
