+++ src/builtin-reporters/summary/markup-printer/utils.ts
