+++ src/reporter/index.ts
@@ -2,6 +2,4 @@
  * Reporter module exports.
  */
 
-export { rawJsonReporter } from "./raw-json-reporter/index.ts";
 export type { Reporter } from "./reporter.ts";
-export { summaryReporter } from "./summary-reporter/index.ts";
